added service layer
