+++ backend/api/routes/login.route.js
@@ -3,7 +3,6 @@ const router = express.Router(); //router do express
 
 // tem que chamar a view Login do frontend para apresentar form de login
 router.post('/login', ); // ??? faz o render e retorna a view login.js
-})
 
 router.post('/logout', )
 module.exports = router;
