+++ request.rest
@@ -2,7 +2,7 @@ GET http://localhost:5000/users
 ###
 GET http://localhost:5000/users/60e57fae3368f31ef8e9ec1f
 ###
-DELETE  http://localhost:5000/users/60e99ad0ffe3c205e891d04a
+DELETE  http://localhost:5000/users/60f18d47c02c342d2cf579e0
 ###
 POST http://localhost:5000/users/add
 Content-Type: application/json
@@ -19,7 +19,7 @@ Content-Type: application/json
 
 {
     "username": "guga.kuerten",
-    "password": "claysurfer",
-    "passcheck": "claysurfer",
-    "isadmin": true
+    "password": "manezinho",
+    "passcheck": "manezinho",
+    "isadmin": false
 }
